Some small updates.
The dragon is now an evil cave caterpillar. Why not.

Additionally bigfoot can no longer jump very high with the "x" command from the ground. Kinda nerfed it, you know.

Refactored how I draw the dragon.

Added in some of the sounds. It's probably not the best way to do it but I'm getting there.